requirements.txt
Django==5.0
djangorestframework==3.15.0
psycopg2==2.9.7
drf-spectacular==0.27.1
django-allauth==0.61.1
python-social-auth==0.3.6
social-auth-app-django==5.4.0
